% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names.R
\name{tract_names}
\alias{tract_names}
\title{Supported XTRACT Tract Names}
\usage{
tract_names
}
\description{
A character vector containing the 42 exact tract name strings
expected by the xtractview package.
}
