% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xtract.R
\name{plot_xtract}
\alias{plot_xtract}
\title{Multi-Planar Anatomical Mapping for XTRACT Data}
\usage{
plot_xtract(tract_data)
}
\arguments{
\item{tract_data}{A data.frame containing columns 'region' (tract names) and 'value'.}
}
\value{
A combined patchwork ggplot object showing 4 anatomical views.
}
\description{
Multi-Planar Anatomical Mapping for XTRACT Data
}
